requirements.txt
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
ta==0.10.2
scipy==1.11.4
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
pytz==2023.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
alpha-vantage==2.3.1
polygon-api-client==1.12.5
ib-insync==0.9.86
py-vollib==1.0.1
quantlib==1.32